tetragon: Move GenericKprobeObjs function to config package
Moving GenericKprobeObjs to config package, because it makes
sense and we get one more bpf package dependency out of
kernels package.
