+++ pkg/kernels/kernels_linux.go
@@ -150,15 +150,3 @@ func IsKernelVersionLessThan(version string) bool {
 	kernelVer, _, _ := GetKernelVersion(option.Config.KernelVersion, option.Config.ProcFS)
 	return (int64(kernelVer) < KernelStringToNumeric(version))
 }
-
-// GenericKprobeObjs returns the generic kprobe and generic retprobe objects
-func GenericKprobeObjs() (string, string) {
-	if EnableV61Progs() {
-		return "bpf_generic_kprobe_v61.o", "bpf_generic_retkprobe_v61.o"
-	} else if MinKernelVersion("5.11") {
-		return "bpf_generic_kprobe_v511.o", "bpf_generic_retkprobe_v511.o"
-	} else if EnableLargeProgs() {
-		return "bpf_generic_kprobe_v53.o", "bpf_generic_retkprobe_v53.o"
-	}
-	return "bpf_generic_kprobe.o", "bpf_generic_retkprobe.o"
-}
